isl_multi_templ.c: extract out isl_multi_dim_id_templ.c
These function are already fairly dubious and they will
make even less sense on an isl_multi_id since the express
purpose of the isl_multi_id is to maintain the dimension
identifiers, shadowing the ephemeral space dimension identifiers.
