+++ Makefile.am
@@ -441,6 +441,7 @@ EXTRA_DIST = \
 	isl_multi_arith_templ.c \
 	isl_multi_cmp.c \
 	isl_multi_coalesce.c \
+	isl_multi_dim_id_templ.c \
 	isl_multi_dims.c \
 	isl_multi_domain_templ.c \
 	isl_multi_floor.c \
