+++ include/isl/multi.h
@@ -22,9 +22,6 @@ __isl_give isl_space *isl_multi_##BASE##_get_space(			\
 __isl_export								\
 __isl_give isl_space *isl_multi_##BASE##_get_domain_space(		\
 	__isl_keep isl_multi_##BASE *multi);				\
-int isl_multi_##BASE##_find_dim_by_name(				\
-	__isl_keep isl_multi_##BASE *multi,				\
-	enum isl_dim_type type, const char *name);			\
 __isl_constructor							\
 __isl_give isl_multi_##BASE *isl_multi_##BASE##_from_##BASE##_list(	\
 	__isl_take isl_space *space, __isl_take isl_##BASE##_list *list); \
@@ -38,18 +35,6 @@ __isl_null isl_multi_##BASE *isl_multi_##BASE##_free(			\
 isl_bool isl_multi_##BASE##_plain_is_equal(				\
 	__isl_keep isl_multi_##BASE *multi1,				\
 	__isl_keep isl_multi_##BASE *multi2);				\
-int isl_multi_##BASE##_find_dim_by_id(					\
-	__isl_keep isl_multi_##BASE *multi, enum isl_dim_type type,	\
-	__isl_keep isl_id *id);						\
-__isl_give isl_id *isl_multi_##BASE##_get_dim_id(			\
-	__isl_keep isl_multi_##BASE *multi,				\
-	enum isl_dim_type type, unsigned pos);				\
-__isl_give isl_multi_##BASE *isl_multi_##BASE##_set_dim_name(		\
-	__isl_take isl_multi_##BASE *multi,				\
-	enum isl_dim_type type, unsigned pos, const char *s);		\
-__isl_give isl_multi_##BASE *isl_multi_##BASE##_set_dim_id(		\
-	__isl_take isl_multi_##BASE *multi,				\
-	enum isl_dim_type type, unsigned pos, __isl_take isl_id *id);	\
 const char *isl_multi_##BASE##_get_tuple_name(				\
 	__isl_keep isl_multi_##BASE *multi, enum isl_dim_type type);	\
 __isl_export								\
@@ -176,6 +161,23 @@ __isl_give isl_multi_##BASE *isl_multi_##BASE##_add_dims(		\
 __isl_give isl_multi_##BASE *						\
 isl_multi_##BASE##_project_domain_on_params(				\
 	__isl_take isl_multi_##BASE *multi);
+
+#define ISL_DECLARE_MULTI_DIM_ID(BASE)					\
+int isl_multi_##BASE##_find_dim_by_name(				\
+	__isl_keep isl_multi_##BASE *multi,				\
+	enum isl_dim_type type, const char *name);			\
+int isl_multi_##BASE##_find_dim_by_id(					\
+	__isl_keep isl_multi_##BASE *multi, enum isl_dim_type type,	\
+	__isl_keep isl_id *id);						\
+__isl_give isl_id *isl_multi_##BASE##_get_dim_id(			\
+	__isl_keep isl_multi_##BASE *multi,				\
+	enum isl_dim_type type, unsigned pos);				\
+__isl_give isl_multi_##BASE *isl_multi_##BASE##_set_dim_name(		\
+	__isl_take isl_multi_##BASE *multi,				\
+	enum isl_dim_type type, unsigned pos, const char *s);		\
+__isl_give isl_multi_##BASE *isl_multi_##BASE##_set_dim_id(		\
+	__isl_take isl_multi_##BASE *multi,				\
+	enum isl_dim_type type, unsigned pos, __isl_take isl_id *id);
 
 #define ISL_DECLARE_MULTI_WITH_DOMAIN(BASE)				\
 __isl_export								\
