+++ isl_aff.c
@@ -3896,6 +3896,7 @@ error:
 #include <isl_multi_apply_set.c>
 #include <isl_multi_arith_templ.c>
 #include <isl_multi_cmp.c>
+#include <isl_multi_dim_id_templ.c>
 #include <isl_multi_dims.c>
 #include <isl_multi_floor.c>
 #include <isl_multi_from_base_templ.c>
@@ -6419,6 +6420,7 @@ error:
 #include <isl_multi_arith_templ.c>
 #include <isl_multi_coalesce.c>
 #include <isl_multi_domain_templ.c>
+#include <isl_multi_dim_id_templ.c>
 #include <isl_multi_dims.c>
 #include <isl_multi_from_base_templ.c>
 #include <isl_multi_gist.c>
@@ -8293,6 +8295,7 @@ error:
 #include <isl_multi_apply_union_set.c>
 #include <isl_multi_arith_templ.c>
 #include <isl_multi_coalesce.c>
+#include <isl_multi_dim_id_templ.c>
 #include <isl_multi_floor.c>
 #include <isl_multi_from_base_templ.c>
 #include <isl_multi_gist.c>
