+++ isl_val.c
@@ -1664,6 +1664,7 @@ isl_stat isl_val_check_match_domain_space(__isl_keep isl_val *v,
 #include <isl_multi_no_explicit_domain.c>
 #include <isl_multi_templ.c>
 #include <isl_multi_arith_templ.c>
+#include <isl_multi_dim_id_templ.c>
 #include <isl_multi_dims.c>
 #include <isl_multi_nan_templ.c>
 #include <isl_multi_product_templ.c>
